feat(event): modify the behaviors of CheckActivityEvent
This event can detects whether the testing has been distracted from the app being tested.
If that happens, we check whether it is caused by the exceptions and crashes via inspecting logcat entries.
If true, testing terminates. If not, we try to return to the app being tested.
We firstly try to press back button to return. If the app does not return, then the app is restarted.

The app is restarted via directly launching the first activity (activity traces are kept during testing),
which does not clean the previous data generated during testing.